refactor: move automatic wildcard session recreation to WASM layer
- Move ensure_wildcard_session_if_expired from SDK to WASM account module
- Only web context should auto-recreate sessions, not SDK directly
- Update execute methods in WASM to check and recreate expired wildcard sessions
- Fix controller test to reflect that SDK no longer auto-recreates sessions
